fix: testing of validation for severity threshold values
