+++ snykTask/src/task-args.ts
@@ -89,17 +89,7 @@ class TaskArgs {
         const errorMsg = `If set, severity threshold must be '${Severity.CRITICAL}' or '${Severity.HIGH}' or '${Severity.MEDIUM}' or '${Severity.LOW}' (case insensitive). If not set, the default is 'low'.`;
         throw new Error(errorMsg);
       } 
-    }
-
-    if (this.severityThreshold && this.failOnThreshold) {
-      const severity = getSeverityOrdinal(this.severityThreshold);
-      const failOn = getSeverityOrdinal(this.failOnThreshold);
-  
-      if (failOn < severity) {
-        const errorMsg = `When both set, fail on threshold must be equal to or greater than severity threshold. ('${this.failOnThreshold}' is less than '${this.severityThreshold}')`;
-        throw new Error(errorMsg);
-      }
-    }    
+    }  
   }
   
   private isNotValidThreshold(threshold: string) {
